flights/requirements.txt
amadeus>=9.0.0
requests>=2.31.0
rich>=13.7.0
python-dotenv>=1.0.0
serpapi>=0.1.5
